Address nullable issues and apply c#9
