+++ src/devices/Amg88xx/Amg88xx.cs
@@ -462,7 +462,7 @@ namespace Iot.Device.Amg88xx
             if (_i2cDevice != null)
             {
                 _i2cDevice?.Dispose();
-                _i2cDevice = null;
+                _i2cDevice = null!;
             }
         }
     }
